feat: add 'not live' mode to disable scheduling
This commit introduces a 'not live' mode to the website, which allows for easily disabling the client scheduling functionality.

This mode is controlled by a 'GoLive' flag in a new configuration file at 'src/config/site.ts'.

When 'GoLive' is set to 'false':
- The Cal.com integration is disabled.
- A message is displayed to users indicating that new clients are not being accepted.

This feature is now documented in the 'README.md' file.